+++ src/panels/ThinkInk_154_Grayscale4_T8.h
@@ -148,13 +148,13 @@ static const uint8_t ti_154t8_gray4_lut_code[] = {
 
 class ThinkInk_154_Grayscale4_T8 : public Adafruit_IL0373 {
 public:
-  ThinkInk_154_Grayscale4_T8(int8_t SID, int8_t SCLK, int8_t DC, int8_t RST,
-                             int8_t CS, int8_t SRCS, int8_t MISO,
-                             int8_t BUSY = -1)
+  ThinkInk_154_Grayscale4_T8(int16_t SID, int16_t SCLK, int16_t DC, int16_t RST,
+                             int16_t CS, int16_t SRCS, int16_t MISO,
+                             int16_t BUSY = -1)
       : Adafruit_IL0373(152, 152, SID, SCLK, DC, RST, CS, SRCS, MISO, BUSY){};
 
-  ThinkInk_154_Grayscale4_T8(int8_t DC, int8_t RST, int8_t CS, int8_t SRCS,
-                             int8_t BUSY = -1, SPIClass *spi = &SPI)
+  ThinkInk_154_Grayscale4_T8(int16_t DC, int16_t RST, int16_t CS, int16_t SRCS,
+                             int16_t BUSY = -1, SPIClass *spi = &SPI)
       : Adafruit_IL0373(152, 152, DC, RST, CS, SRCS, BUSY, spi){};
 
   void begin(thinkinkmode_t mode = THINKINK_GRAYSCALE4) {
