+++ src/panels/ThinkInk_290_Grayscale4_T5.h
@@ -148,13 +148,13 @@ const uint8_t ti_290t5_gray4_lut_code[] = {
 
 class ThinkInk_290_Grayscale4_T5 : public Adafruit_IL0373 {
 public:
-  ThinkInk_290_Grayscale4_T5(int8_t SID, int8_t SCLK, int8_t DC, int8_t RST,
-                             int8_t CS, int8_t SRCS, int8_t MISO,
-                             int8_t BUSY = -1)
+  ThinkInk_290_Grayscale4_T5(int16_t SID, int16_t SCLK, int16_t DC, int16_t RST,
+                             int16_t CS, int16_t SRCS, int16_t MISO,
+                             int16_t BUSY = -1)
       : Adafruit_IL0373(296, 128, SID, SCLK, DC, RST, CS, SRCS, MISO, BUSY){};
 
-  ThinkInk_290_Grayscale4_T5(int8_t DC, int8_t RST, int8_t CS, int8_t SRCS,
-                             int8_t BUSY = -1, SPIClass *spi = &SPI)
+  ThinkInk_290_Grayscale4_T5(int16_t DC, int16_t RST, int16_t CS, int16_t SRCS,
+                             int16_t BUSY = -1, SPIClass *spi = &SPI)
       : Adafruit_IL0373(296, 128, DC, RST, CS, SRCS, BUSY, spi){};
 
   void begin(thinkinkmode_t mode = THINKINK_GRAYSCALE4) {
